refactor(matrix): move inbound dlq code into submodule
Owning abstraction: Matrix inbound-DLQ persistence, replay, quarantine, encrypted record codecs, and store-rekey recovery live in channels::matrix::inbound_dlq.

Invariant: encrypted DLQ envelope versions, legacy refusal policy, replay merge/cap behavior, quarantine durability, and rekey backup semantics preserve their existing on-disk and runtime behavior after the move.

Validation: /Users/zera/build/carapace/scripts/cargo-serial fmt --check; /Users/zera/build/carapace/scripts/cargo-serial check --all-targets; /Users/zera/build/carapace/scripts/cargo-serial nextest run dlq.
